GUACAMOLE-764: Merge correction to size of RDP file read/write offsets (all should be 64-bit).
(cherry picked from commit dc71987fcaa3e5f37e97ae7224fffa3b933f15f5)

+++ src/protocols/rdp/fs.c
@@ -403,7 +403,7 @@ int guac_rdp_fs_open(guac_rdp_fs* fs, const char* path,
 
 }
 
-int guac_rdp_fs_read(guac_rdp_fs* fs, int file_id, int offset,
+int guac_rdp_fs_read(guac_rdp_fs* fs, int file_id, uint64_t offset,
         void* buffer, int length) {
 
     int bytes_read;
@@ -427,7 +427,7 @@ int guac_rdp_fs_read(guac_rdp_fs* fs, int file_id, int offset,
 
 }
 
-int guac_rdp_fs_write(guac_rdp_fs* fs, int file_id, int offset,
+int guac_rdp_fs_write(guac_rdp_fs* fs, int file_id, uint64_t offset,
         void* buffer, int length) {
 
     int bytes_written;
